fix: tweak code for 9269cf35 and add remove handle.
